backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.9
sse-starlette>=1.8.0
starlette>=0.36.0
gitpython>=3.1.40
aiofiles>=23.2.1

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
alembic>=1.13.0

# Celery & Async Processing
celery>=5.3.0
flower>=2.0.0
django-celery-beat>=2.5.0

# HTTP Client for API calls
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
cryptography>=42.0.0

# Stripe Payment Processing
stripe>=8.0.0

# Object Storage (S3/MinIO)
boto3>=1.34.0
minio>=7.2.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=24.1.0
psutil>=5.9.0

# Rate Limiting
slowapi>=0.1.9
limits>=3.8.0

# Utilities
email-validator>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
# Admin Auth
firebase-admin>=6.5.0

# AI / LLM Providers
google-genai>=1.0.0
openai>=1.0.0
anthropic>=0.40.0

# Neo4j Graph Database (Attack Path Intelligence)
neo4j>=5.0.0
